LayoutTests/inspector/unit-tests/object-expected.txt

Testing WebInspector.Object.


== Running test suite: Object
-- Running test case: Events propagation
Dispatch count: 1
